feat: start 함수 구현
